Allow for customization of shell buffer names
This change adds a new customizable variable, called
kubel-shell-buffer-name-format. This variable defines the buffer name format
for shell, ansi-term, eshell, and vterm buffers created through
kubel-exec-popup. I frequently have to interact with multiple clusters and
multiple namespaces in each cluster, so having the ability to add the
Kubernetes context and namespace to the buffer name is useful. As explained in
the documentation for kubel-shell-buffer-name-format, the following %-sequences
are available:

%t: The shell type. Examples of this are `shell` and `eshell`
%c: The container name
%p: The pod name
%n: The current namespace
%C: The current context

I set the default format to "kubel - %t - %c@%p", which means that the buffer
names for shell, eshell, and vterm buffers are not changed by
default. ansi-term buffers currently use a different format for the buffer
name, so this code change does change the name of ansi-term buffers by default.